pyproject.toml
[tool.poetry]
name = "cdc-decrypt"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "cdc_decrypt"}]

[tool.poetry.dependencies]
python = "^3.11"
pycryptodome = "^3.17"
pandas = "^2.0.1"
pytest = "^7.3.1"
coverage = "^7.2.5"
pyarrow = "^12.0.0"
fsspec = "^2023.5.0"
s3fs = "^2023.5.0"
boto3 = "^1.17"
line-profiler = "^4.0.3"


[tool.poetry.group.dev.dependencies]
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
